+++ DuckDuckGo.Fluent.Plugin/DuckResultFactory.cs
