+++ DuckDuckGo.Fluent.Plugin/ResultType.cs
